Move ExoClient to shared location
