+++ bench/eval_tool_calls.py
@@ -15,7 +15,7 @@ from pathlib import Path
 from typing import Any, Literal
 
 import httpx
-from harness import (
+from exo.client import (
     ExoClient,
     ExoHttpError,
     add_common_instance_args,
