+++ bench/exo_bench.py
@@ -30,7 +30,7 @@ from pathlib import Path
 from statistics import mean
 from typing import Any
 
-from harness import (
+from exo.client import (
     ExoClient,
     ExoHttpError,
     add_common_instance_args,
